IFRS_Backend/requirements.txt
# Django + DRF (Exact matches from local)
Django==5.2.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.2
python-decouple==3.8
dj-database-url==2.3.0
whitenoise==6.6.0

# Cloudinary Media
django-cloudinary-storage==0.3.0
cloudinary==1.40.0
pillow==11.1.0

# PostgreSQL
psycopg2-binary==2.9.10

# Celery Stack
celery==5.5.3
redis==5.3.0
flower==2.0.1

# External APIs & Data Processing
requests==2.32.3
types-requests==2.32.4.20250611
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.4.10
twilio==9.6.1

# Server
gunicorn==23.0.0

# AI & Pydantic
pydantic==2.11.4
pydantic_core==2.33.2
google-genai==1.2.0
google-cloud-texttospeech>=2.0.0

# Web3
web3==7.12.1
eth-account==0.13.7